requirements.txt
   flask
   azure-identity
   azure-keyvault-secrets
   azure-storage-blob